+++ lib/widgets/viewer/action/video_action_delegate.dart
@@ -140,12 +140,15 @@ class VideoActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAwareMix
   Future<void> _showStreamSelectionDialog(BuildContext context, AvesVideoController controller) async {
     final streams = controller.streams;
     final currentSelectedStreams = await Future.wait(MediaStreamType.values.map(controller.getSelectedStream));
-    final currentSelectedIndices = currentSelectedStreams.whereNotNull().map((v) => v.index).toSet();
 
     final userSelectedStreams = await showDialog<Map<MediaStreamType, MediaStreamSummary?>>(
       context: context,
       builder: (context) => VideoStreamSelectionDialog(
-        streams: Map.fromEntries(streams.map((stream) => MapEntry(stream, currentSelectedIndices.contains(stream.index)))),
+        streams: Map.fromEntries(streams.map((stream) {
+          final selectedStream = currentSelectedStreams.whereNotNull().firstWhereOrNull((v) => v.type == stream.type);
+          final selected = selectedStream != null && selectedStream.index == stream.index;
+          return MapEntry(stream, selected);
+        })),
       ),
       routeSettings: const RouteSettings(name: VideoStreamSelectionDialog.routeName),
     );