+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+## <a id="v1.9.5"></a>[v1.9.5] - 2023-09-17
+
+### Fixed
+
+- crash when cataloguing some videos
+- workflow when moving to an album with insufficient storage
+
 ## <a id="v1.9.4"></a>[v1.9.4] - 2023-09-13
 
 ### Changed